moved model config and added registration extension
