+++ InMeal/InMeal.Core/InMeal.Core.csproj
@@ -24,7 +24,6 @@
 
     <ItemGroup>
         <Folder Include="Migrations" />
-        <Folder Include="ModelConfiguration\" />
     </ItemGroup>
 
     <ItemGroup>
