+++ InMeal/InMeal.Core/ModelConfiguration/RecipeConfig.cs
@@ -1,8 +1,9 @@
-﻿using InMeal.Core.Globalisation.Generators;
+﻿using InMeal.Core.Entities;
+using InMeal.Core.Globalisation.Generators;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace InMeal.Core.Entities;
+namespace InMeal.Core.ModelConfiguration;
 
 public class RecipeConfig
 {
